Make sketch.pde work on Processing 3
size() can not be called inside setup() if it the dimensions are specified in variables.

Also, drawing inside an oscEvent() seems to have no effect (no image is displayed). I assume it's because oscEvent() runs in a separate thread, which is not the graphics thread, so no drawing is allowed.

These changes have been tested on Linux 64 bit, Processing 3.0.2.

The order of servers specified on issue #114 had to be also modified to make the example work.